+++ Vault/src/utopia/vault/database/Store.scala
@@ -1,14 +1,14 @@
 package utopia.vault.database
 
 import utopia.flow.collection.CollectionExtensions._
-import utopia.flow.collection.immutable.{IntSet, OptimizedIndexedSeq, Pair}
+import utopia.flow.collection.immutable.{IntSet, OptimizedIndexedSeq}
 import utopia.flow.collection.template.MapAccess
 import utopia.flow.operator.Identity
 import utopia.flow.operator.MaybeEmpty.collectionMayBeEmpty
 import utopia.flow.view.immutable.caching.Lazy
 import utopia.flow.view.template.Extender
 import utopia.vault.database.Store.ReplaceHandler
-import utopia.vault.database.Store.ReplaceHandler.MappingReplaceHandler
+import utopia.vault.database.Store.ReplaceHandler.{DeprecatingReplaceHandler, MappingReplaceHandler}
 import utopia.vault.nosql.view.{TimeDeprecatableView, ViewManyByIntIds}
 import utopia.vault.store.{HasId, StoreResult}
 
@@ -29,17 +29,18 @@ object Store
 	 * @tparam S Type of stored instances
 	 * @return A factory for constructing store interfaces which use the specified model
 	 */
-	def apply[D, S <: HasId[Int] with Extender[D]](model: Inserter[D, S]) = new PreparedStoreFactory[D, S](model)
+	def apply[D, S <: HasId[Int] with Extender[D]](model: Inserter[D, S]) = new StoreFactory[D, S](model)
 	
 	/**
 	 * @param model Interface used for inserting new items to the database
 	 * @param toData A function which converts the accepted input into insertable data
-	 * @tparam V Type of accepted input
+	 * @tparam In Type of accepted input
 	 * @tparam D Type of stored data
 	 * @tparam S Type of the items that have already been stored to the DB
 	 * @return An interface for storing items to the DB
 	 */
-	def using[V, D, S <: HasId[Int]](model: Inserter[D, S])(toData: V => D) = new PreparedStore[V, D, S](model)(toData)
+	def using[In, D, S <: HasId[Int]](model: Inserter[D, S])(toData: In => D) =
+		new PreparedStore[In, D, S](model)(toData)
 	/**
 	 * @param model Interface used for inserting new items to the database
 	 * @tparam D Type of stored data
@@ -59,10 +60,22 @@ object Store
 	 def keyMappedUsing[K, V, D, S <: HasId[Int]](model: Inserter[D, S])(toData: (K, V) => D) =
 		 new PreparedKeyMappedStore[K, V, D, S](model, None)(toData)
 	
+	/**
+	 * Prepares an advanced store factory
+	 * @param model DB model to use in data inserts
+	 * @param toData A function which converts input to insertable data
+	 * @tparam In Type of accepted input / items-to-store
+	 * @tparam D Type of inserted data
+	 * @tparam S Type of stored items (after insert)
+	 * @return A factory for creating the store interface
+	 */
+	def advancedUsing[In, D, S](model: Inserter[D, S])(toData: In => D) =
+		new AdvancedStoreFactory[In, D, S](model)(toData)
+	
 	
 	// NESTED   ---------------------------
 	
-	class PreparedStoreFactory[D, S <: HasId[Int] with Extender[D]](model: Inserter[D, S])
+	class StoreFactory[D, S <: HasId[Int] with Extender[D]](model: Inserter[D, S])
 	{
 		// COMPUTED ---------------------------
 		
@@ -76,10 +89,10 @@ object Store
 		
 		/**
 		 * @param toData A function which converts the accepted input into insertable data
-		 * @tparam V Type of accepted input
+		 * @tparam In Type of accepted input
 		 * @return An interface for storing items to the DB
 		 */
-		def apply[V](toData: V => D) = new PreparedStore[V, D, S](model)(toData)
+		def apply[In](toData: In => D) = new PreparedStore[In, D, S](model)(toData)
 		/**
 		 * @param toData A function which converts the accepted input (key + value) into insertable data
 		 * @tparam K Type of keys attached to the stored values
@@ -89,15 +102,22 @@ object Store
 		def keyMapped[K, V](toData: (K, V) => D) = new PreparedKeyMappedStore[K, V, D, S](model, None)(toData)
 		
 		/**
+		 * @param toData A function which converts accepted input into insertable data
+		 * @tparam In Type of accepted input / items-to-store
+		 * @return A factory for constructing more complex store interfaces
+		 */
+		def advanced[In](toData: In => D) = new AdvancedStoreFactory[In, D, S](model)(toData)
+		
+		/**
 		 * @param toData A function for converting values into data.
 		 *               Accepts:
 		 *                  1. Value to convert
 		 *                  1. A store construction parameter
-		 * @tparam V Type of the converted values
-		 * @tparam I Type of the additional construction parameter
+		 * @tparam In Type of the converted values
+		 * @tparam GN Type of the additional construction parameter
 		 * @return A new factory for creating store interfaces, using construction parameters
 		 */
-		def generic[V, I](toData: (V, I) => D) = new GenericStoreFactory[V, D, S, I](model, None)(toData)
+		def generic[In, GN](toData: (In, GN) => D) = new GenericStoreFactory[In, D, S, GN](model, None)(toData)
 		/**
 		 * @param toData A function for converting key-value pairs into data.
 		 *               Accepts:
@@ -106,14 +126,50 @@ object Store
 		 *                  1. A store construction parameter
 		 * @tparam K Type of the unique keys used
 		 * @tparam V Type of the converted values
-		 * @tparam I Type of the additional construction parameter
+		 * @tparam GN Type of the additional construction parameter
 		 * @return A new factory for creating store interfaces, using construction parameters
 		 */
-		def genericKeyMapped[K, V, I](toData: (K, V, I) => D) =
-			new GenericKeyMapStoreFactory[K, V, D, S, I](model, None)(toData)
+		def genericKeyMapped[K, V, GN](toData: (K, V, GN) => D) =
+			new GenericKeyMapStoreFactory[K, V, D, S, GN](model, None)(toData)
 	}
 	
-	trait GenericStoreFactoryLike[+V, +S <: HasId[Int], -I, +A, +Repr] extends CanUseReplaceHandler[V, S, Repr]
+	class AdvancedStoreFactory[In, -D, S](model: Inserter[D, S])(toData: In => D)
+	{
+		/**
+		 * Creates a store interface, which wraps inserted values
+		 * @param wrapInserted A function which converts an inserted item
+		 *                     to a type comparable with that of the existing items
+		 * @tparam E Type of the accepted existing items
+		 * @return A new store interface
+		 */
+		//noinspection ConvertibleToMethodValue
+		def mapInserted[E <: HasId[Int]](wrapInserted: S => E) =
+			apply[E, StoreResult[E]] { i => StoreResult.inserted(wrapInserted(i)) } { StoreResult.existed(_) }
+		/**
+		 * Creates a store interface, which wraps existing values to another type
+		 * @param wrapExisting A function which converts a pre-existing item to a stored item
+		 * @tparam E Type of the accepted existing items
+		 * @return A new store interface
+		 */
+		def mapExisting[E](wrapExisting: E => S)(implicit ev: S <:< HasId[Int]) =
+			apply[E, StoreResult[S]] { i => StoreResult(i, i.id, isNew = true) } { e =>
+				val wrapped = wrapExisting(e)
+				StoreResult(wrapped, wrapped.id, isNew = false)
+			}
+		
+		/**
+		 * Creates a new custom-mapping store
+		 * @param insertedToResult A function which converts an inserted item into a result
+		 * @param existingToResult A function which converts a pre-existing item into a result
+		 * @tparam E Type of the accepted existing items
+		 * @tparam R Type of generated store results
+		 * @return A new store interface
+		 */
+		def apply[E, R](insertedToResult: S => R)(existingToResult: E => R) =
+			new AdvancedPreparedStore[In, E, D, S, R](model)(toData)(insertedToResult)(existingToResult)
+	}
+	
+	trait GenericStoreFactoryLike[+In, +E, +S, -GN, +ST, +Repr] extends CanUseReplaceHandler[In, E, S, Repr]
 	{
 		// ABSTRACT    ----------------------------
 		
@@ -121,7 +177,7 @@ object Store
 		 * @param param A store construction parameter
 		 * @return A store interface, based on that parameter
 		 */
-		def apply(param: I): A
+		def apply(param: GN): ST
 		
 		
 		// COMPUTED -------------------------------
@@ -130,46 +186,59 @@ object Store
 		 * @param param A store construction parameter (implicit)
 		 * @return A store interface, based on the implicit parameter
 		 */
-		def contextual(implicit param: I) = apply(param)
+		def contextual(implicit param: GN) = apply(param)
 	}
-	class GenericStoreFactory[V, -D, S <: HasId[Int], -I](model: Inserter[D, S],
-	                                                    replaceHandler: Option[ReplaceHandler[V, S]])
-	                                                   (toData: (V, I) => D)
-		extends GenericStoreFactoryLike[V, S, I, PreparedStore[V, D, S], GenericStoreFactory[V, D, S, I]]
+	class GenericStoreFactory[In, -D, S <: HasId[Int], -GN](model: Inserter[D, S],
+	                                                        replaceHandler: Option[ReplaceHandler[In, S, S]])
+	                                                       (toData: (In, GN) => D)
+		extends GenericStoreFactoryLike[In, S, S, GN, PreparedStore[In, D, S], GenericStoreFactory[In, D, S, GN]]
 	{
 		// IMPLEMENTED  ---------------------------
 		
-		override def apply(param: I) =
-			new PreparedStore[V, D, S](model, replaceHandler)({ value => toData(value, param) })
+		override def apply(param: GN) =
+			new PreparedStore[In, D, S](model, replaceHandler)({ value => toData(value, param) })
 		
-		override def using(handler: ReplaceHandler[V, S]): GenericStoreFactory[V, D, S, I] =
-			new GenericStoreFactory[V, D, S, I](model, Some(handler))(toData)
+		override def using(handler: ReplaceHandler[In, S, S]): GenericStoreFactory[In, D, S, GN] =
+			new GenericStoreFactory[In, D, S, GN](model, Some(handler))(toData)
 	}
-	class GenericKeyMapStoreFactory[K, V, -D, S <: HasId[Int], -I](model: Inserter[D, S],
-	                                                             replaceHandler: Option[ReplaceHandler[(K, V), S]])
-	                                                            (toData: (K, V, I) => D)
-		extends GenericStoreFactoryLike[(K, V), S, I, PreparedKeyMappedStore[K, V, D, S], GenericKeyMapStoreFactory[K, V, D, S, I]]
+	class GenericKeyMapStoreFactory[K, V, -D, S <: HasId[Int], -GN](model: Inserter[D, S],
+	                                                                replaceHandler: Option[ReplaceHandler[(K, V), S, S]])
+	                                                               (toData: (K, V, GN) => D)
+		extends GenericStoreFactoryLike[(K, V), S, S, GN, PreparedKeyMappedStore[K, V, D, S], GenericKeyMapStoreFactory[K, V, D, S, GN]]
 	{
 		// IMPLEMENTED  ----------------------------
 		
-		override def apply(param: I): PreparedKeyMappedStore[K, V, D, S] =
+		override def apply(param: GN): PreparedKeyMappedStore[K, V, D, S] =
 			new PreparedKeyMappedStore[K, V, D, S](model, replaceHandler)({ (key, value) => toData(key, value, param) })
 		
-		override def using(handler: ReplaceHandler[(K, V), S]): GenericKeyMapStoreFactory[K, V, D, S, I] =
-			new GenericKeyMapStoreFactory[K, V, D, S, I](model, Some(handler))(toData)
+		override def using(handler: ReplaceHandler[(K, V), S, S]): GenericKeyMapStoreFactory[K, V, D, S, GN] =
+			new GenericKeyMapStoreFactory[K, V, D, S, GN](model, Some(handler))(toData)
 	}
 	
-	class PreparedStore[V, -D, S <: HasId[Int]](model: Inserter[D, S],
-	                                            replaceHandler: Option[ReplaceHandler[V, S]] = None)
-	                                           (toData: V => D)
-		extends Store[V, S, PreparedStore[V, D, S]]
+	abstract class AbstractStore[In, E, -D, S, +R, +Repr](model: Inserter[D, S],
+	                                                      replaceHandler: Option[ReplaceHandler[In, E, S]] = None)
+	                                                     (toData: In => D)
+		extends Store[In, E, S, R, Repr]
 	{
+		// ABSTRACT -------------------------------
+		
+		/**
+		 * Wraps an inserted item as a store result
+		 * @param inserted The inserted item
+		 * @return A store result wrapping the inserted item
+		 */
+		protected def insertedResult(inserted: S): R
+		/**
+		 * Wraps an existing item as a store result
+		 * @param existing The existing item
+		 * @return A store result wrapping the specified item
+		 */
+		protected def existingResult(existing: E): R
+		
+		
 		// IMPLEMENTED  ---------------------------
 		
-		override def using(handler: ReplaceHandler[V, S]): PreparedStore[V, D, S] =
-			new PreparedStore(model, Some(handler))(toData)
-		
-		override def single(item: V, existingMatch: Option[S])(implicit connection: Connection) =
+		override def single(item: In, existingMatch: Option[E])(implicit connection: Connection) =
 			existingMatch match {
 				// Case: There already existed a new item => Performs the replacement check, if appropriate
 				case Some(existing) =>
@@ -180,30 +249,30 @@ object Store
 							if (replacer.handleMatch(item, item, existing)) {
 								val lazyInserted = Lazy { model.insert(toData(item)) }
 								replacer.replace(MapAccess { _ => lazyInserted.value })
-								StoreResult.inserted(lazyInserted.value)
+								insertedResult(lazyInserted.value)
 							}
 							// Case: Duplicate => Yields the existing item
 							else
-								StoreResult.existed(existing)
+								existingResult(existing)
 						
 						// Case: Replacement is not used
 						//       => Considers the new item a duplicate and yields the existing item
-						case None => StoreResult.existed(existing)
+						case None => existingResult(existing)
 					}
 				// Case: No existing match => Inserts the item
-				case None => StoreResult.inserted(model.insert(toData(item)))
+				case None => insertedResult(model.insert(toData(item)))
 			}
 		
 		//noinspection ConvertibleToMethodValue
-		override def keyMapped[K](itemsToStore: IterableOnce[(K, V)], existingItems: Map[K, S])
-		                (implicit connection: Connection) =
+		override def keyMapped[K](itemsToStore: IterableOnce[(K, In)], existingItems: Map[K, E])
+		                         (implicit connection: Connection) =
 		{
-			val existingResultView = existingItems.view.mapValues { StoreResult.existed(_) }
+			val existingResultView = existingItems.view.mapValues(existingResult)
 			itemsToStore.nonEmptyIterator match {
 				case Some(itemsIterator) =>
 					// Checks against the existing items,
 					// preparing the non-duplicates as new inserts and handling possible replacements
-					val insertsBuilder = OptimizedIndexedSeq.newBuilder[(K, V)]
+					val insertsBuilder = OptimizedIndexedSeq.newBuilder[(K, In)]
 					itemsIterator.foreach { case (key, item) =>
 						existingItems.get(key) match {
 							// Case: Matches an existing item
@@ -230,7 +299,7 @@ object Store
 							// Performs the replacement, if appropriate
 							replaceHandler.foreach { replacer =>
 								val lazyInsertedMap = lazyInserted.map { _.toMap[Any, S] }
-								replacer.replace(MapAccess.wrapLazily[Any, S](lazyInsertedMap.value))
+								replacer.replace(MapAccess.wrap[Any, S](lazyInsertedMap))
 							}
 							// Converts the insertion results into store-results
 							// If the insert was not performed yet, it is performed here
@@ -238,26 +307,47 @@ object Store
 							View.concat(
 								existingResultView,
 								lazyInserted.value.view.map { case (key, inserted) =>
-									key -> StoreResult.inserted(inserted)
+									key -> insertedResult(inserted)
 								}
 							).toMap
 						
 						// Case: None of the specified items were new => Returns the existing items
 						case None => existingResultView.toMap
 					}
-					
+				
 				// Case: There were no items to store => Yields the existing items
 				case None => existingResultView.toMap
 			}
 		}
 	}
+	
+	abstract class AbstractSimpleStore[In, -D, S <: HasId[Int], +Repr](model: Inserter[D, S],
+	                                                                   replaceHandler: Option[ReplaceHandler[In, S, S]] = None)
+	                                                                  (toData: In => D)
+		extends AbstractStore[In, S, D, S, StoreResult[S], Repr](model, replaceHandler)(toData)
+	{
+		override protected def insertedResult(inserted: S) = StoreResult.inserted(inserted)
+		override protected def existingResult(existing: S) = StoreResult.existed(existing)
+	}
+	
+	class PreparedStore[In, -D, S <: HasId[Int]](model: Inserter[D, S],
+	                                             replaceHandler: Option[ReplaceHandler[In, S, S]] = None)
+	                                            (toData: In => D)
+		extends AbstractSimpleStore[In, D, S, PreparedStore[In, D, S]](model, replaceHandler)(toData)
+	{
+		// IMPLEMENTED  ---------------------------
+		
+		override def using(handler: ReplaceHandler[In, S, S]): PreparedStore[In, D, S] =
+			new PreparedStore[In, D, S](model, Some(handler))(toData)
+	}
+	
 	class PreparedStoreData[D, S <: Extender[D] with HasId[Int]](model: Inserter[D, S],
-	                                                             replaceHandler: Option[ReplaceHandler[D, S]] = None)
-		extends PreparedStore[D, D, S](model, replaceHandler)(Identity) with Store[D, S, PreparedStoreData[D, S]]
+	                                                             replaceHandler: Option[ReplaceHandler[D, S, S]] = None)
+		extends AbstractSimpleStore[D, D, S, PreparedStoreData[D, S]](model, replaceHandler)(Identity)
 	{
 		// IMPLEMENTED  ------------------------
 		
-		override def using(handler: ReplaceHandler[D, S]): PreparedStoreData[D, S] =
+		override def using(handler: ReplaceHandler[D, S, S]): PreparedStoreData[D, S] =
 			new PreparedStoreData[D, S](model, Some(handler))
 		
 		
@@ -276,59 +366,20 @@ object Store
 		             (implicit connection: Connection) =
 			super.keyMap[K](itemsToStore, existingItems) { dataToKey(_) } { e => dataToKey(e.wrapped) }
 	}
+
 	class PreparedKeyMappedStore[K, V, -D, S <: HasId[Int]](model: Inserter[D, S],
-	                                                        replaceHandler: Option[ReplaceHandler[(K, V), S]])
+	                                                        replaceHandler: Option[ReplaceHandler[(K, V), S, S]])
 	                                                       (toData: (K, V) => D)
-		extends PreparedStore[(K, V), D, S](model, replaceHandler)({ case (key, value) => toData(key, value) })
-			with Store[(K, V), S, PreparedKeyMappedStore[K, V, D, S]]
+		extends AbstractSimpleStore[(K, V), D, S, PreparedKeyMappedStore[K, V, D, S]](model, replaceHandler)(
+			{ case (key, value) => toData(key, value) })
 	{
 		// IMPLEMENTED  ---------------------
 		
-		override def using(handler: ReplaceHandler[(K, V), S]): PreparedKeyMappedStore[K, V, D, S] =
+		override def using(handler: ReplaceHandler[(K, V), S, S]): PreparedKeyMappedStore[K, V, D, S] =
 			new PreparedKeyMappedStore[K, V, D, S](model, Some(handler))(toData)
 			
 		
 		// OTHER    -------------------------
-		
-		/**
-		 * @param rootAccess Root-level access to the matching database entries.
-		 *                   Used for deprecating older item versions. Supports timestamp-based deprecation.
-		 * @param shouldReplace A function that accepts the proposed item (value) and the item that already exists in the DB.
-		 *                          - Yields true if the new item should be considered a new version
-		 *                            and replace the existing item.
-		 *                          - Yields false if the new item should be considered a duplicate
-		 *                            and ignored / not inserted to the DB.
-		 * @return Copy of this interface, which deprecates older item versions by specifying a deprecation timestamp
-		 */
-		def deprecatingValues(rootAccess: ViewManyByIntIds[TimeDeprecatableView[_]])(shouldReplace: (V, S) => Boolean) =
-			using(ReplaceHandler.deprecating(rootAccess)(shouldReplace).mapInput { _._2 })
-		/**
-		 * @param shouldReplace A function that accepts the proposed item (value) and the item that already exists in the DB.
-		 *                          - Yields true if the new item should be considered a new version
-		 *                            and replace the existing item.
-		 *                          - Yields false if the new item should be considered a duplicate
-		 *                            and ignored / not inserted to the DB.
-		 * @param replace A function that performs the actual replacement, based on two parameters:
-		 *                      1. Replacements to perform. Never empty.
-		 *                         Each entry is a pair containing two stored versions:
-		 *                              1. The newly inserted version
-		 *                              1. The earlier version, which needs to be replaced / deprecated
-		 *                      1. Database connection to use
-		 * @return Copy of this interface, which deprecates older item versions by using the specified replace function
-		 */
-		def replacingValues(shouldReplace: (V, S) => Boolean)(replace: (IndexedSeq[Pair[S]], Connection) => Unit) =
-			using(ReplaceHandler(shouldReplace)(replace).mapInput { _._2 })
-		
-		/**
-		 * @param itemsToStore Items to store, including unique keys
-		 * @param existingItems Existing DB entries
-		 * @param existingToKey A function which converts an existing item into a unique key
-		 * @param connection Implicit DB connection
-		 * @return A map where each encountered key is mapped to the stored item (either existing or inserted)
-		 */
-		def apply(itemsToStore: IterableOnce[(K, V)], existingItems: IterableOnce[S])(existingToKey: S => K)
-		         (implicit connection: Connection) =
-			keyMap(itemsToStore, existingItems) { _._1 }(existingToKey)
 		
 		/**
 		 * Stores items after mapping them to keys and values
@@ -344,9 +395,29 @@ object Store
 		def keyValueMap[A](itemsToStore: IterableOnce[A], existingItems: IterableOnce[S])
 		                  (itemToKey: A => K)(itemToValue: A => V)(existingToKey: S => K)
 		                  (implicit connection: Connection) =
-		{
 			apply(itemsToStore.iterator.map { a => itemToKey(a) -> itemToValue(a) }, existingItems)(existingToKey)
-		}
+		/**
+		 * @param itemsToStore Items to store, including unique keys
+		 * @param existingItems Existing DB entries
+		 * @param existingToKey A function which converts an existing item into a unique key
+		 * @param connection Implicit DB connection
+		 * @return A map where each encountered key is mapped to the stored item (either existing or inserted)
+		 */
+		def apply(itemsToStore: IterableOnce[(K, V)], existingItems: IterableOnce[S])(existingToKey: S => K)
+		         (implicit connection: Connection) =
+			keyMap(itemsToStore, existingItems) { _._1 }(existingToKey)
+	}
+	
+	class AdvancedPreparedStore[In, E, -D, S, +R](model: Inserter[D, S],
+	                                              replaceHandler: Option[ReplaceHandler[In, E, S]] = None)
+	                                             (toData: In => D)(wrapInsert: S => R)(wrapExisting: E => R)
+		extends AbstractStore[In, E, D, S, R, AdvancedPreparedStore[In, E, D, S, R]](model, replaceHandler)(toData)
+	{
+		override protected def insertedResult(inserted: S): R = wrapInsert(inserted)
+		override protected def existingResult(existing: E): R = wrapExisting(existing)
+		
+		override def using(handler: ReplaceHandler[In, E, S]): AdvancedPreparedStore[In, E, D, S, R] =
+			new AdvancedPreparedStore[In, E, D, S, R](model, Some(handler))(toData)(wrapInsert)(wrapExisting)
 	}
 	
 	object ReplaceHandler
@@ -361,13 +432,13 @@ object Store
 		 *                            and replace the existing item.
 		 *                          - Yields false if the new item should be considered a duplicate
 		 *                            and ignored / not inserted to the DB.
-		 * @tparam V Type of new values
-		 * @tparam S Type of existing DB entries
+		 * @tparam In Type of new values
+		 * @tparam E Type of existing DB entries
 		 * @return A new replace-handler, which deprecates by updating a timestamp column
 		 */
-		def deprecating[V, S <: HasId[Int]](rootAccess: ViewManyByIntIds[TimeDeprecatableView[_]])
-		                                   (shouldReplace: (V, S) => Boolean) =
-			new DeprecatingReplaceHandler[V, S](rootAccess)(shouldReplace)
+		def deprecating[In, E <: HasId[Int]](rootAccess: ViewManyByIntIds[TimeDeprecatableView[_]])
+		                                    (shouldReplace: (In, E) => Boolean) =
+			DeprecatingReplaceHandler[In, E](rootAccess)(shouldReplace)
 		
 		/**
 		 * Creates a new replace-handler, where the replacements are performed after the inserts
@@ -382,20 +453,34 @@ object Store
 		 *                              1. The newly inserted version
 		 *                              1. The earlier version, which needs to be replaced / deprecated
 		 *                      1. Database connection to use
-		 * @tparam V Type of new items, before they've been inserted
-		 * @tparam S Type of items after they've been inserted
+		 * @tparam In Type of new items, before they've been inserted
+		 * @tparam E Type of the existing items
+		 * @tparam S Type of stored / inserted items
 		 * @return A new replace-handler
 		 */
-		def apply[V, S](shouldReplace: (V, S) => Boolean)
-		               (replace: (IndexedSeq[Pair[S]], Connection) => Unit): ReplaceHandler[V, S] =
-			new _ReplaceHandler[V, S](shouldReplace)(replace)
+		def apply[In, E, S](shouldReplace: (In, E) => Boolean)
+		                   (replace: (IndexedSeq[(S, E)], Connection) => Unit): ReplaceHandler[In, E, S] =
+			new _ReplaceHandler[In, E, S](shouldReplace)(replace)
 		
 		
 		// NESTED   -------------------------
 		
-		class DeprecatingReplaceHandler[-V, -S <: HasId[Int]](rootAccess: ViewManyByIntIds[TimeDeprecatableView[_]])
-		                                                     (shouldReplace: (V, S) => Boolean)
-			extends ReplaceHandler[V, S]
+		object DeprecatingReplaceHandler
+		{
+			/**
+			 * @param rootAccess Root-level access point to items which may be deprecated
+			 * @param shouldReplace A function for testing whether a new item (1) should replace an existing item (2).
+			 * @tparam In Type of evaluated new items
+			 * @tparam E Type of existing items
+			 * @return A new replace-handler
+			 */
+			def apply[In, E <: HasId[Int]](rootAccess: ViewManyByIntIds[TimeDeprecatableView[_]])
+			                              (shouldReplace: (In, E) => Boolean) =
+				new DeprecatingReplaceHandler[In, E](rootAccess)(shouldReplace)(_.id)
+		}
+		class DeprecatingReplaceHandler[-In, -E](rootAccess: ViewManyByIntIds[TimeDeprecatableView[_]])
+		                                        (shouldReplace: (In, E) => Boolean)(idOf: E => Int)
+			extends ReplaceHandler[In, E, Any]
 		{
 			// ATTRIBUTES   -----------------
 			
@@ -405,36 +490,36 @@ object Store
 			
 			// IMPLEMENTED  -----------------
 			
-			override def handleMatch(key: Any, newItem: V, existingItem: S): Boolean = {
+			override def handleMatch(key: Any, newItem: In, existingItem: E): Boolean = {
 				// Case: Replacement => Remembers the old version's ID
 				if (shouldReplace(newItem, existingItem)) {
-					idsToReplaceBuilder += existingItem.id
+					idsToReplaceBuilder += idOf(existingItem)
 					true
 				}
 				else
 					false
 			}
 			
-			override def replace(inserted: MapAccess[Any, S])(implicit connection: Connection): Unit = {
+			override def replace(inserted: MapAccess[Any, Any])(implicit connection: Connection): Unit = {
 				// Deprecates the rows matching the collected IDs
 				rootAccess(idsToReplaceBuilder.result()).deprecate()
 				idsToReplaceBuilder.clear()
 			}
 		}
 		
-		private class _ReplaceHandler[-V, S](shouldReplace: (V, S) => Boolean)
-		                                   (replace: (IndexedSeq[Pair[S]], Connection) => Unit)
-			extends ReplaceHandler[V, S]
+		private class _ReplaceHandler[-In, E, S](shouldReplace: (In, E) => Boolean)
+		                                        (replace: (IndexedSeq[(S, E)], Connection) => Unit)
+			extends ReplaceHandler[In, E, S]
 		{
 			// ATTRIBUTES   -----------------------
 			
 			// Collects the items to replace, mapped to their unique keys
-			private val replacementsBuilder = OptimizedIndexedSeq.newBuilder[(Any, S)]
+			private val replacementsBuilder = OptimizedIndexedSeq.newBuilder[(Any, E)]
 			
 			
 			// IMPLEMENTED  -----------------------
 			
-			override def handleMatch(key: Any, newItem: V, existingItem: S): Boolean = {
+			override def handleMatch(key: Any, newItem: In, existingItem: E): Boolean = {
 				// Case: Replacement => Remembers the old version, as well as its key
 				if (shouldReplace(newItem, existingItem)) {
 					replacementsBuilder += (key -> existingItem)
@@ -446,15 +531,15 @@ object Store
 			
 			override def replace(inserted: MapAccess[Any, S])(implicit connection: Connection): Unit = {
 				// Performs the replacements. Safely assumes that at least one replacement was prepared.
-				replace(replacementsBuilder.result().map { case (key, existing) => Pair(inserted(key), existing) },
+				replace(replacementsBuilder.result().map { case (key, existing) => inserted(key) -> existing },
 					connection)
 				replacementsBuilder.clear()
 			}
 		}
-		private class MappingReplaceHandler[-I, V, S](delegate: ReplaceHandler[V, S], f: I => V)
-			extends ReplaceHandler[I, S]
+		private class MappingReplaceHandler[-In, M, E, -S](delegate: ReplaceHandler[M, E, S], f: In => M)
+			extends ReplaceHandler[In, E, S]
 		{
-			override def handleMatch(key: Any, newItem: I, existingItem: S): Boolean =
+			override def handleMatch(key: Any, newItem: In, existingItem: E): Boolean =
 				delegate.handleMatch(key, f(newItem), existingItem)
 			
 			override def replace(inserted: MapAccess[Any, S])(implicit connection: Connection): Unit =
@@ -467,10 +552,10 @@ object Store
 	 * ReplaceHandlers are expected to be stateful,
 	 * but to reset their state after the replace operation has been performed.
 	 *
-	 * @tparam V Type of new items
-	 * @tparam S Type of existing items
+	 * @tparam In Type of new items
+	 * @tparam E Type of existing items
 	 */
-	trait ReplaceHandler[-V, -S]
+	trait ReplaceHandler[-In, -E, -S]
 	{
 		// ABSTRACT --------------------------
 		
@@ -483,7 +568,7 @@ object Store
 		 * @return Whether 'newItem' should be considered an updated version of 'existingItem'.
 		 *         False if 'newItem' should be considered a duplicate and ignored.
 		 */
-		def handleMatch(key: Any, newItem: V, existingItem: S): Boolean
+		def handleMatch(key: Any, newItem: In, existingItem: E): Boolean
 		
 		/**
 		 * Performs the prepared replacements.
@@ -505,14 +590,20 @@ object Store
 		
 		/**
 		 * @param f A mapping function applied to the stored items, before they're passed to this handler
-		 * @tparam V2 Type of mapping results
+		 * @tparam I2 Type of mapping results
 		 * @return A copy of this handler, which accepts the items before mapping them
 		 */
-		def mapInput[V2](f: V2 => V): ReplaceHandler[V2, S] = new MappingReplaceHandler[V2, V, S](this, f)
+		def mapInput[I2](f: I2 => In): ReplaceHandler[I2, E, S] = new MappingReplaceHandler[I2, In, E, S](this, f)
 	}
 }
 
-trait CanUseReplaceHandler[+V, +S <: HasId[Int], +Repr]
+/**
+ * Common trait for store interfaces which support the application of replace-handlers
+ * @tparam In Type of accepted input
+ * @tparam E Type of the existing (potentially replaced) items
+ * @tparam Repr Implementing type
+ */
+trait CanUseReplaceHandler[+In, +E, +S, +Repr]
 {
 	// ABSTRACT ------------------------
 	
@@ -520,7 +611,7 @@ trait CanUseReplaceHandler[+V, +S <: HasId[Int], +Repr]
 	 * @param handler A new replace-handler to perform replacements with
 	 * @return A copy of this interface, which uses the specified replace-handler
 	 */
-	def using(handler: ReplaceHandler[V, S]): Repr
+	def using(handler: ReplaceHandler[In, E, S]): Repr
 	
 	
 	// OTHER    -----------------------
@@ -535,8 +626,9 @@ trait CanUseReplaceHandler[+V, +S <: HasId[Int], +Repr]
 	 *                            and ignored / not inserted to the DB.
 	 * @return Copy of this interface, which deprecates older item versions by specifying a deprecation timestamp
 	 */
-	def deprecating(rootAccess: ViewManyByIntIds[TimeDeprecatableView[_]])(shouldReplace: (V, S) => Boolean) =
-		using(ReplaceHandler.deprecating(rootAccess)(shouldReplace))
+	def deprecating(rootAccess: ViewManyByIntIds[TimeDeprecatableView[_]])(shouldReplace: (In, E) => Boolean)
+	               (implicit ev: E <:< HasId[Int]) =
+		using(new DeprecatingReplaceHandler(rootAccess)(shouldReplace)(_.id))
 	/**
 	 * @param shouldReplace A function that accepts the proposed item and the item that already exists in the DB.
 	 *                          - Yields true if the new item should be considered a new version
@@ -551,17 +643,55 @@ trait CanUseReplaceHandler[+V, +S <: HasId[Int], +Repr]
 	 *                      1. Database connection to use
 	 * @return Copy of this interface, which deprecates older item versions by using the specified replace function
 	 */
-	def replacing(shouldReplace: (V, S) => Boolean)(replace: (IndexedSeq[Pair[S]], Connection) => Unit) =
+	def replacing(shouldReplace: (In, E) => Boolean)(replace: (IndexedSeq[(S, E)], Connection) => Unit) =
 		using(ReplaceHandler(shouldReplace)(replace))
+	
+	
+	// OTHER    -------------------------------
+	
+	/**
+	 * @param rootAccess Root-level access to the matching database entries.
+	 *                   Used for deprecating older item versions. Supports timestamp-based deprecation.
+	 * @param shouldReplace A function that accepts the proposed item (value) and the item that already exists in the DB.
+	 *                          - Yields true if the new item should be considered a new version
+	 *                            and replace the existing item.
+	 *                          - Yields false if the new item should be considered a duplicate
+	 *                            and ignored / not inserted to the DB.
+	 * @return Copy of this interface, which deprecates older item versions by specifying a deprecation timestamp
+	 */
+	def deprecatingValues[Val](rootAccess: ViewManyByIntIds[TimeDeprecatableView[_]])
+	                          (shouldReplace: (Val, E) => Boolean)
+	                          (implicit evIn: In <:< (_, Val), evEx: E <:< HasId[Int]) =
+		using(new DeprecatingReplaceHandler(rootAccess)(shouldReplace)(_.id).mapInput { _._2 })
+	/**
+	 * @param shouldReplace A function that accepts the proposed item (value) and the item that already exists in the DB.
+	 *                          - Yields true if the new item should be considered a new version
+	 *                            and replace the existing item.
+	 *                          - Yields false if the new item should be considered a duplicate
+	 *                            and ignored / not inserted to the DB.
+	 * @param replace A function that performs the actual replacement, based on two parameters:
+	 *                      1. Replacements to perform. Never empty.
+	 *                         Each entry is a pair containing two stored versions:
+	 *                              1. The newly inserted version
+	 *                              1. The earlier version, which needs to be replaced / deprecated
+	 *                      1. Database connection to use
+	 * @tparam Val Type of the key-mapped input values
+	 * @return Copy of this interface, which deprecates older item versions by using the specified replace function
+	 */
+	def replacingValues[Val](shouldReplace: (Val, E) => Boolean)(replace: (IndexedSeq[(S, E)], Connection) => Unit)
+	                        (implicit ev: In <:< (_, Val)) =
+		using(ReplaceHandler(shouldReplace)(replace).mapInput { _._2 })
 }
 
 /**
  * A prepared interface for performing various store functions
- * @tparam V Type of accepted input / items to store
- * @tparam S Type of existing database entries / inserted items
+ * @tparam In Type of accepted input / items to store
+ * @tparam E Type of existing items
+ * @tparam S Type of inserted items
+ * @tparam R Type of store results
  * @tparam Repr Type of this store interface
  */
-trait Store[V, S <: HasId[Int], +Repr] extends CanUseReplaceHandler[V, S, Repr]
+trait Store[In, E, +S, +R, +Repr] extends CanUseReplaceHandler[In, E, S, Repr]
 {
 	// ABSTRACT ------------------------
 	
@@ -573,7 +703,7 @@ trait Store[V, S <: HasId[Int], +Repr] extends CanUseReplaceHandler[V, S, Repr]
 	 * @param connection Implicit DB connection
 	 * @return Result of this store operation, either containing the inserted entry, or 'existingMatch'.
 	 */
-	def single(item: V, existingMatch: Option[S])(implicit connection: Connection): StoreResult[S]
+	def single(item: In, existingMatch: Option[E])(implicit connection: Connection): R
 	
 	/**
 	 * Stores 0-n items to the database. Checks against existing data and won't insert any duplicate entries.
@@ -586,8 +716,8 @@ trait Store[V, S <: HasId[Int], +Repr] extends CanUseReplaceHandler[V, S, Repr]
 	 * @return A map where keys are the specified keys and values are store results,
 	 *         either containing a newly inserted item, or one of the existing DB entries.
 	 */
-	def keyMapped[K](itemsToStore: IterableOnce[(K, V)], existingItems: Map[K, S])
-	                (implicit connection: Connection): Map[K, StoreResult[S]]
+	def keyMapped[K](itemsToStore: IterableOnce[(K, In)], existingItems: Map[K, E])
+	                (implicit connection: Connection): Map[K, R]
 	
 	
 	// OTHER    ------------------------
@@ -600,7 +730,7 @@ trait Store[V, S <: HasId[Int], +Repr] extends CanUseReplaceHandler[V, S, Repr]
 	 * @param connection Implicit DB connection
 	 * @return Result of this store operation, either containing the inserted entry, or 'existingMatch'.
 	 */
-	def single(item: V, existingMatch: S)(implicit connection: Connection): StoreResult[S] =
+	def single(item: In, existingMatch: E)(implicit connection: Connection): R =
 		single(item, Some(existingMatch))
 	
 	/**
@@ -615,8 +745,8 @@ trait Store[V, S <: HasId[Int], +Repr] extends CanUseReplaceHandler[V, S, Repr]
 	 * @return A map where keys are the specified keys and values are store results,
 	 *         either containing a newly inserted item, or one of the existing DB entries.
 	 */
-	def keyMap[K](itemsToStore: IterableOnce[V], existingItems: IterableOnce[S])
-	             (itemToKey: V => K)(existingToKey: S => K)(implicit connection: Connection) =
+	def keyMap[K](itemsToStore: IterableOnce[In], existingItems: IterableOnce[E])
+	             (itemToKey: In => K)(existingToKey: E => K)(implicit connection: Connection) =
 		keyMapped[K](itemsToStore.iterator.map { item => itemToKey(item) -> item },
 			existingItems.iterator.map { existing => existingToKey(existing) -> existing }.toMap)
 	
@@ -626,14 +756,14 @@ trait Store[V, S <: HasId[Int], +Repr] extends CanUseReplaceHandler[V, S, Repr]
 	 * @param connection Implicit DB connection
 	 * @return The inserted item
 	 */
-	def unique(item: V)(implicit connection: Connection) = single(item, None)
+	def unique(item: In)(implicit connection: Connection) = single(item, None)
 	/**
 	 * Inserts 0-n items to the database. Assumes that no matching entries exist in the database.
 	 * @param items The items to insert
 	 * @param connection Implicit DB connection
 	 * @return The inserted items (not necessarily in the same order as specified).
 	 */
-	def unique(items: IterableOnce[V])(implicit connection: Connection) =
+	def unique(items: IterableOnce[In])(implicit connection: Connection) =
 		keyMapUnique(items)(Identity).values
 	
 	/**
@@ -644,7 +774,7 @@ trait Store[V, S <: HasId[Int], +Repr] extends CanUseReplaceHandler[V, S, Repr]
 	 * @param connection Implicit DB connection
 	 * @return A map containing the inserted items, each mapped to a key produced with 'toKey'
 	 */
-	def keyMapUnique[K](items: IterableOnce[V])(toKey: V => K)(implicit connection: Connection) =
+	def keyMapUnique[K](items: IterableOnce[In])(toKey: In => K)(implicit connection: Connection) =
 		keyMappedUnique(items.iterator.map { i => toKey(i) -> i })
 	/**
 	 * Inserts 0-n key-mapped items to the database. Assumes that no matching entries exist in the database.
@@ -652,6 +782,17 @@ trait Store[V, S <: HasId[Int], +Repr] extends CanUseReplaceHandler[V, S, Repr]
 	 * @param connection Implicit DB connection
 	 * @return A map containing the inserted items, each mapped to their respective key.
 	 */
-	def keyMappedUnique[K](items: IterableOnce[(K, V)])(implicit connection: Connection) =
-		keyMapped(items, Map.empty[K, S])
+	def keyMappedUnique[K](items: IterableOnce[(K, In)])(implicit connection: Connection) =
+		keyMapped(items, Map.empty[K, E])
+	
+	/**
+	 * @param itemsToStore Items to store, including unique keys
+	 * @param existingItems Existing DB entries
+	 * @param existingToKey A function which converts an existing item into a unique key
+	 * @param connection Implicit DB connection
+	 * @return A map where each encountered key is mapped to the stored item (either existing or inserted)
+	 */
+	def keyMapExisting[K, V](itemsToStore: IterableOnce[In], existingItems: IterableOnce[E])(existingToKey: E => K)
+	                        (implicit connection: Connection, ev: In <:< (K, V)) =
+		keyMap(itemsToStore, existingItems) { _._1 }(existingToKey)
 }